login/api_views.py
from rest_framework import generics
from .models import Usuario
from .serializers import UsuarioSerializer

class UsuarioListAPI(generics.ListCreateAPIView):
    queryset = Usuario.objects.all()
    serializer_class = UsuarioSerializer

class UsuarioDetailAPI(generics.RetrieveUpdateDestroyAPIView):
    queryset = Usuario.objects.all()
    serializer_class = UsuarioSerializer